+++ pdf-generator/functions/src/utilities/event_channel.ts
@@ -1,8 +1,7 @@
-import * as admin from "firebase-admin";
-
-admin.initializeApp();
-
 import { getEventarc } from "firebase-admin/eventarc";
+
+export const EVENT_TYPE_PREFIX =
+  "firebase.extensions.pdfplum.pdf-generator.v1.";
 
 export const eventChannel = process.env.EVENTARC_CHANNEL
   ? getEventarc().channel(process.env.EVENTARC_CHANNEL, {